move autoIRAD to benchmark
