+++ benchmark/autoIRAD/images/.gitignore
@@ -1,2 +1,1 @@
 /images.zip
-*.png
