+++ pyproject.toml
@@ -31,6 +31,12 @@ where = ["."]
 include = ["sageml*"]
 
 [dependency-groups]
+benchmark = [
+    "h2o>=3.46.0.7",
+    "opencv-python>=4.12.0.88",
+    "tensorflow>=2.20.0",
+    "tpot>=1.1.0",
+]
 dev = [
     "click>=8.1.8",
     "pydataset>=0.2.0",
